requirements.txt
asgiref==3.2.10
Django==3.1.1
django-rest-framework==0.1.0
djangorestframework==3.12.1
gunicorn==20.0.4
numpy==1.19.1
opencv-python==4.4.0.42
Pillow==7.2.0
pytesseract==0.3.6
pytz==2020.1
pyzbar==0.1.8
sqlparse==0.3.1
opencv-python-headless==4.4.0.44
requests==2.24.0
cloudinary==1.22.0
whitenoise==5.2.0
celery==5.0.2
